+++ src/engine.cpp
@@ -4,7 +4,6 @@ Author: Alexander Mastryukov
 License: http://creativecommons.org/licenses/by-nc-nd/4.0/legalcode
 */
 
-#include "item.h"
 #include "engine.h"
 
 Engine::Engine()
@@ -40,7 +39,7 @@ void Engine::init(std::string ID)
   Json::Value items = value["engines"];
 
   //go through the json file and find the item by ID
-  for (int i = 0; i < items.size(); i++)
+  for (size_t i = 0; i < items.size(); i++)
   {
     if (ID.compare(items[i]["id"].asCString()) == 0)
     {
